pyproject.toml
[tool.poetry]
name = "plexy"
version = "0.1.0"
description = "A CLI-based AI assistant with tool-calling capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src", from = "."}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.0"
openai = "^1.3.0"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
plexy = "src.main:plexy"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 